minor fixes for Fahad's test
