+++ Pipfile
@@ -1,0 +1,36 @@
+[[source]]
+name = "pypi"
+url = "https://pypi.org/simple"
+verify_ssl = true
+
+[dev-packages]
+
+[packages]
+aniso8601 = "==8.0.0"
+astroid = "==2.3.3"
+autopep8 = "==1.5"
+click = "==7.0"
+contextlib2 = "==0.5.5"
+gunicorn = "==20.0.4"
+isort = "==4.3.21"
+itsdangerous = "==1.1.0"
+lazy-object-proxy = "==1.4.3"
+mccabe = "==0.6.1"
+ortools = "==7.5.7466"
+protobuf = "==3.11.2"
+pycodestyle = "==2.5.0"
+pylint = "==2.4.4"
+python-dotenv = "==0.10.5"
+pytz = "==2019.3"
+schema = "==0.7.1"
+six = "==1.14.0"
+typed-ast = "==1.4.1"
+wrapt = "==1.11.2"
+Flask = "==1.1.1"
+Flask-RESTful = "==0.3.7"
+Jinja2 = "==2.11.1"
+MarkupSafe = "==1.1.1"
+Werkzeug = "==0.16.1"
+
+[requires]
+python_version = "3.7"
